Migrate to Next.js: App Router, SEO and metadata
