+++ package.json
@@ -4,37 +4,31 @@
   "version": "2.1.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "lint": "eslint .",
-    "preview": "vite preview"
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   },
   "dependencies": {
-    "@gsap/react": "^2.1.2",
-    "@radix-ui/react-slot": "^1.2.3",
-    "@tailwindcss/vite": "^4.0.4",
     "@vercel/analytics": "^1.5.0",
-    "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "dependencies": "^0.0.1",
     "framer-motion": "^12.23.12",
+    "next": "^16.1.4",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-icons": "^5.5.0",
-    "react-router-dom": "^7.6.1",
-    "tailwind-merge": "^3.3.1",
-    "tailwindcss": "^4.0.4"
+    "tailwind-merge": "^3.3.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.19.0",
     "@types/react": "^19.0.8",
     "@types/react-dom": "^19.0.3",
-    "@vitejs/plugin-react": "^4.3.4",
+    "autoprefixer": "^10.4.23",
     "eslint": "^9.19.0",
     "eslint-plugin-react": "^7.37.4",
     "eslint-plugin-react-hooks": "^5.0.0",
-    "eslint-plugin-react-refresh": "^0.4.18",
     "globals": "^15.14.0",
-    "vite": "^6.1.0"
+    "postcss": "^8.5.6",
+    "tailwindcss": "^3.4.1"
   }
 }
