+++ src/app/globals.css
@@ -1,0 +1,95 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+:root {
+    --color-bg: #1a1a1a;
+    /* Fallback/Converted from oklch(10% 0 0) */
+    --color-surface: #1f1f1f;
+    --color-border: #333333;
+
+    --color-primary: #B9FF66;
+    /* Neon Lime */
+    --color-secondary: #ff00ff;
+    /* Hot Pink */
+    --color-white: #ffffff;
+    --color-muted: #a3a3a3;
+
+    /* Fonts */
+    --font-sans: "Inter", "Satoshi", system-ui, sans-serif;
+    --font-mono: "JetBrains Mono", ui-monospace;
+}
+
+html,
+body {
+    font-family: var(--font-sans);
+    scroll-behavior: smooth;
+    background-color: #111;
+    /* Fallback background */
+    color: #fff;
+}
+
+/* Custom Utilities */
+@layer utilities {
+    .perspective-1000 {
+        perspective: 1000px;
+    }
+
+    .preserve-3d {
+        transform-style: preserve-3d;
+    }
+
+    .backface-hidden {
+        backface-visibility: hidden;
+    }
+}
+
+/* Animations */
+@keyframes slideIn {
+    from {
+        opacity: 0;
+        transform: translateY(30px);
+    }
+
+    to {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
+
+.animate-slideIn {
+    animation: slideIn 1s ease-out forwards;
+}
+
+@keyframes slideWipe {
+    0% {
+        transform: translateX(100%);
+        opacity: 0;
+    }
+
+    100% {
+        transform: translateX(0%);
+        opacity: 1;
+    }
+}
+
+.slide-enter {
+    animation: slideWipe 1s ease-in-out forwards;
+}
+
+@keyframes fade-in-up {
+    from {
+        opacity: 0;
+        transform: translateY(30px);
+    }
+
+    to {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
+
+.animate-fade-in-up {
+    animation: fade-in-up 0.6s ease-out forwards;
+    opacity: 0;
+}
