+++ src/components/Header.jsx
@@ -1,3 +1,4 @@
+'use client';
 import { useState, useEffect } from "react";
 
 export default function Header() {
