+++ src/data/projects.js
@@ -1,0 +1,63 @@
+export const projects = [
+    {
+        title: "Design and Development of  Automated Household Waste Segregation System",
+        category: "ANSYS",
+        image: "/images/household-waste.png"
+    },
+    {
+        title: "ML Based Predictive Modeling of Wire EDM Performance Metrics",
+        category: "AI/ML",
+        image: "/images/ML-Predictive.jpeg"
+    },
+    {
+        title: "CFD-Based Aerodynamic Analysis of the NACA 0012 Airfoil",
+        category: "ANSYS",
+        image: "/images/air-foil.png"
+    },
+    {
+        title: "BioSkin-AI: A Machine Learning-Powered Self-Healing Prosthetic Skin",
+        category: "AI/ML",
+        image: "/images/srijan.png"
+    },
+    {
+        title: "Computer vision-driven Teleoperation Robotic Hand",
+        category: "AI/ML",
+        image: "/images/sushil.png"
+    },
+    {
+        title: "Fish Fin-Inspired Heat Sink Design",
+        category: "OpenSCAD/CAD",
+        image: "/images/heat-sink.png"
+    },
+    {
+        title: "Wack-a-Mole Game",
+        category: "UI/UX Game",
+        image: "/images/Wack-a-mole.jpg"
+    },
+    {
+        title: "Traffic Light Simulation",
+        category: "Embedded Systems / IoT Simulation",
+        image: "/images/trafficLight.jpg"
+    },
+    {
+        title: "Melody Buzzers",
+        category: "Embedded Systems / Sound Engineering",
+        image: "/images/buzzer.png",
+    },
+    {
+        title: "Stack Tower Game",
+        category: "UI/UX Design",
+        image: "/images/stack-tower.png"
+    },
+    {
+        title: "RGB Light Control",
+        category: "Embedded Systems / Electronics Simulation",
+        image: "..//images/rgb-light1.png"
+    },
+    {
+        title: "Reverse Countdown Display",
+        category: "Product Design",
+        image: "..//images/countdown.png"
+    },
+    { title: 'ECM CGPA Tracker', category: 'UI/UX', image: '..//images/cgpaTracker.png' },
+];
