+++ src/sections/Contact.jsx
@@ -1,3 +1,4 @@
+'use client';
 import { useRef, useState } from 'react';
 import { motion } from 'framer-motion';
